+++ web/packages/gallery/services/file.ts
@@ -6,9 +6,13 @@ import { apiURL } from "@/base/origins";
 import type {
     EncryptedMagicMetadata,
     EnteFile,
+    FilePublicMagicMetadata,
+    FilePublicMagicMetadataProps,
     FileWithUpdatedMagicMetadata,
     FileWithUpdatedPublicMagicMetadata,
 } from "@/media/file";
+import { mergeMetadata } from "@/media/file";
+import { updateMagicMetadata } from "@/new/photos/services/magic-metadata";
 import HTTPService from "@ente/shared/network/HTTPService";
 import { getToken } from "@ente/shared/storage/localStorage/helpers";
 
@@ -115,3 +119,55 @@ export const updateFilePublicMagicMetadata = async (
         }),
     );
 };
+
+export async function changeFileName(
+    file: EnteFile,
+    editedName: string,
+): Promise<EnteFile> {
+    const updatedPublicMagicMetadataProps: FilePublicMagicMetadataProps = {
+        editedName,
+    };
+
+    const updatedPublicMagicMetadata: FilePublicMagicMetadata =
+        await updateMagicMetadata(
+            updatedPublicMagicMetadataProps,
+            file.pubMagicMetadata,
+            file.key,
+        );
+    const updateResult = await updateFilePublicMagicMetadata([
+        { file, updatedPublicMagicMetadata },
+    ]);
+    // @ts-ignore
+    return updateResult[0];
+}
+
+export async function changeCaption(
+    file: EnteFile,
+    caption: string,
+): Promise<EnteFile> {
+    const updatedPublicMagicMetadataProps: FilePublicMagicMetadataProps = {
+        caption,
+    };
+
+    const updatedPublicMagicMetadata: FilePublicMagicMetadata =
+        await updateMagicMetadata(
+            updatedPublicMagicMetadataProps,
+            file.pubMagicMetadata,
+            file.key,
+        );
+    const updateResult = await updateFilePublicMagicMetadata([
+        { file, updatedPublicMagicMetadata },
+    ]);
+    // @ts-ignore
+    return updateResult[0];
+}
+
+export function updateExistingFilePubMetadata(
+    existingFile: EnteFile,
+    updatedFile: EnteFile,
+) {
+    // @ts-ignore
+    existingFile.pubMagicMetadata = updatedFile.pubMagicMetadata;
+    // @ts-ignore
+    existingFile.metadata = mergeMetadata([existingFile])[0].metadata;
+}
